fix: move logs related files to separate folder
