+++ package.json
@@ -20,7 +20,7 @@
     }
   ],
   "scripts": {
-    "compare": "ts-node src/test/utils/compare-logs.ts",
+    "compare": "ts-node src/logs/compare-logs.ts",
     "build:fast": "yarn clean && yarn gen:grammar && yarn gen:stdlib && yarn gen:func-js && tsc --project tsconfig.fast.json && yarn copy:stdlib && yarn copy:func && yarn to-relative",
     "build": "cross-env NODE_OPTIONS=--max_old_space_size=4096 tsc && yarn copy:stdlib && yarn copy:func && yarn to-relative",
     "gen:config": "ts-to-zod -k --skipValidation src/config/config.ts src/config/config.zod.ts",
