+++ src/logs/parse-log.ts
@@ -5,8 +5,8 @@ import { parse as parseYaml } from "yaml";
 import { z } from "zod";
 import { parse, str, type Loc as ParserLoc } from "@tonstudio/parser-runtime";
 import { relative } from "path";
-import type { $ast } from "@/test/utils/logs";
-import * as $ from "@/test/utils/logs";
+import type { $ast } from "@/logs/logs";
+import * as $ from "@/logs/logs";
 import { getSrcInfo } from "@/grammar/src-info";
 
 const packageSchema = z.array(
